Allow setting of uninitalised properties
The get_object_vars() method only returns initialised properties. It is
perfectly legitimate to instantiate an object without setting all
properties initially and to then update in a subsequent call.

Sadly this is not possible using get_object_vars() so I have modified
the approach to extend upon the existing Reflection.